refactor(ai): 共有ビルドを build-bundle + bundle-sink に分離（本番プレビュー基盤）
本番ライブプレビュー(方式D/A)の共通土台。ビルドを「成果物(Buffer配列)化」と
「配置先(sink)」に分離し、同じ vite build 経路をローカルディスクと本番 Blob の
両方に流用できるようにした。

- build-bundle.ts: buildBundle(code)→BundleFile[] （ディスクに残さず Buffer で返す）
- bundle-sink.ts: BundleSink 抽象 + localDiskSink（slug検証/atomic rename/traversal防止）
  + getBundleSink() 選択点（本番 Blob sink は次段でここに分岐）
- local-build.ts: buildSharedBundle/readSharedAsset/removeSharedBundle を上記へ委譲（公開API不変）

検証: ローカルで公開→ビルド→.ai-shared配置→/s-assets配信(200)→/s/<slug>描画(200)、
不正slug/traversal=404 を確認（挙動・セキュリティとも従来同等）。
